Switch to visualizer being an optional feature to build depends
