+++ Cargo.toml
@@ -6,13 +6,15 @@ rust-version = "1.76"
 default-run = "visualizer"
 
 [dependencies]
-clap = { version = "4.5.28", features = ["derive"] }
 itertools = "0.14.0"
-random_color = "1.0.0"
-rerun = "0.21.0"
 serde = { version = "1.0.217", features = ["derive"] }
 serde_json = "1.0.138"
 walkdir = "2.5.0"
+
+clap = { version = "4.5.28", features = ["derive"], optional = true }
+random_color = { version = "1.0.0", optional = true }
+rerun = { version = "0.22.0", optional = true }
+
 
 [dev-dependencies]
 assert_approx_eq = "1.1.0"
@@ -22,6 +24,9 @@ rstest = "0.24.0"
 rstest_reuse = "*"
 tempfile = "3.16.0"
 
+[features]
+visualizer = ["clap", "random_color", "rerun"]
+
 [[bin]]
 name = "generate-rotated-ipa-polygons"
 path = "src/bin/generate_rotated_ipa_polygons.rs"
@@ -29,6 +34,7 @@ path = "src/bin/generate_rotated_ipa_polygons.rs"
 [[bin]]
 name = "visualizer"
 path = "src/bin/run_visualizer.rs"
+required-features = ["visualizer"]
 
 [[bench]]
 name = "extreme_points"
